bug fixed

+++ credit-sys/src/main/java/com/taoyes3/credit/sys/constant/Constant.java
@@ -6,7 +6,7 @@ package com.taoyes3.credit.sys.constant;
  */
 public class Constant {
     // 超级管理员ID
-    public static final int SUPER_ADMIN_ID = 1;
+    public static final Long SUPER_ADMIN_ID = 1L;
     // 系统菜单最大id
-    public static final int SYS_MENU_MAX_ID = 1;
+    public static final Long SYS_MENU_MAX_ID = 1L;
 }
